Refactor shared helpers into utils module and update docs
